[Feature] Adds a view for creating new projects
Introduces a new view that allows users to input a project name and save it.

This view includes a text field for the project name and a button to trigger the save action. Currently, the save action is a placeholder and will be implemented later to persist the project to SwiftData.